Initial psABI extensions for MIPS.
The Relocations section is definitely a placeholder, and some of the other
portions could use additional detail.

Some of this duplicates content in the CHERI programmer's guide.  We
should probably only document these in one place.  We should arguably
move some things such as how var-args works into this document.