Add most of a CHERI-RISC-V ELF psABI specification
The only notable omission is of the calling convention itself, being the
most awkward to specify (and not implemented quite how I'd like; in
particular with respect to pairs of capabilities). That is, however, one
of the less-important bits for anyone wishing to get a feel for what the
code generation is doing, and is also fairly obvious to work out for
primitive data types; the complexity only arises once aggregates
(especially unions) are involved.